requirements.txt
streamlit
supabase
pandas
plotly
google-generativeai